+++ src/global.d.ts
@@ -1,5 +1,4 @@
 declare const MICROSOFT_CLARITY_ID: string;
-declare const MELLODDY: boolean;
 declare const DEFAULT_PAGE_SIZE: number;
 declare const __APP_VERSION__: string;
 declare const API_URL: string;