+++ src/modules/series/SeriesTypes.ts
@@ -12,7 +12,6 @@ export interface SerieFeaturesT {
 
 export interface PointT {
     rank: number;
-    epoch: number;
     round: number;
     perf: number | null;
     testTaskKey: string | null;