+++ src/modules/series/SeriesUtils.ts
@@ -50,10 +50,10 @@ function findSerie(
     return null;
 }
 
-function getEpochOrRoundAsNumber(str: string | null): number {
+function getRoundAsNumber(str: string | null): number {
     /**
-     * epoch & round may not be present
-     * epoch & round are stored as string in the db
+     * round may not be present
+     * round is stored as string in the db
      */
     try {
         if (typeof str === 'string') {
@@ -61,49 +61,6 @@ function getEpochOrRoundAsNumber(str: string | null): number {
         } else return -1;
     } catch {
         return -1;
-    }
-}
-
-function fixEpoch(serie: SerieT) {
-    /**
-     * MDY testtuples have been registered with faulty epochs numbers.
-     * This is because the first and second testtuples are about the same epoch:
-     *  - rank 0 is about the beginning of epoch 0
-     *  - next rank is about the end of epoch 0
-     *
-     * Testtuple registration code has been changed but is faulty even after the change.
-     * Here are the situations for the first 3 testtuples:
-     *  - before 2022-01-26
-     *    Rank | Epoch | Correct Epoch
-     *    ---- | ----- | -------------
-     *       0 |     0 |             0
-     *      98 |     0 |             1
-     *     198 |     1 |             2
-     *
-     *  - after 2022-01-26:
-     *    Rank | Epoch | Correct Epoch
-     *    ---- | ----- | -------------
-     *       0 |     1 |             0
-     *      98 |     1 |             1
-     *     198 |     2 |             2
-     *
-     * NB: 98/198 are ranks for multi partner CP, other types of CP have different ranks.
-     * We cannot use the rank value to infer epoch, only the order.
-     */
-
-    const firstPoint = serie.points[0];
-    if (firstPoint.rank === 0 && firstPoint.epoch === 0) {
-        // before 2022-01-26 situation:
-        // the rank 0 epoch is the only correct one, all other ones must be incremented by 1
-        for (const point of serie.points) {
-            if (point.rank !== 0) {
-                point.epoch += 1;
-            }
-        }
-    } else if (firstPoint.rank === 0 && firstPoint.epoch === 1) {
-        // after 2022-01-26 situation:
-        // the rank 0 epoch is the only wrong one and must be set to 0
-        firstPoint.epoch = 0;
     }
 }
 
@@ -119,8 +76,7 @@ export function buildSeries(
             rank: performance.compute_task.rank,
             perf: performance.perf,
             testTaskKey: performance.compute_task.key,
-            epoch: getEpochOrRoundAsNumber(performance.compute_task.epoch),
-            round: getEpochOrRoundAsNumber(performance.compute_task.round_idx),
+            round: getRoundAsNumber(performance.compute_task.round_idx),
         };
 
         const serieFeatures = buildSerieFeatures(performance, computePlanKey);
@@ -137,11 +93,6 @@ export function buildSeries(
                 ...serieFeatures,
             });
         }
-    }
-
-    // In MDY, testtuple registration didn't use the correct epochs
-    for (const serie of series) {
-        fixEpoch(serie);
     }
 
     return series;
@@ -161,67 +112,6 @@ export function buildSeriesGroups(series: SerieT[]): SerieT[][] {
         groups.push(metricGroup);
     }
     return groups;
-}
-
-const average = (values: number[]): number => {
-    return values.reduce((sum, value) => sum + value, 0) / values.length;
-};
-
-export function buildAverageSerie(
-    series: SerieT[],
-    name: string
-): SerieT | null {
-    if (series.length < 2) {
-        return null;
-    }
-
-    const ranksPerfs: Record<
-        number,
-        { epoch: number; round: number; perfs: number[] }
-    > = {};
-
-    for (const serie of series) {
-        for (const point of serie.points) {
-            if (point.perf !== null) {
-                if (ranksPerfs[point.rank]) {
-                    ranksPerfs[point.rank] = {
-                        epoch: point.epoch,
-                        round: point.round,
-                        perfs: [...ranksPerfs[point.rank].perfs, point.perf],
-                    };
-                } else {
-                    ranksPerfs[point.rank] = {
-                        epoch: point.epoch,
-                        round: point.round,
-                        perfs: [point.perf],
-                    };
-                }
-            }
-        }
-    }
-
-    // only includes points where we have values for all series
-    const points: PointT[] = Object.entries(ranksPerfs)
-        .filter(([, { perfs }]) => perfs.length === series.length)
-        .map(([rank, { epoch, round, perfs }]) => ({
-            rank: parseInt(rank),
-            epoch,
-            round,
-            perf: average(perfs),
-            testTaskKey: null,
-        }));
-
-    return {
-        id: uuidv4(),
-        points: points,
-        algoKey: uuidv4(),
-        datasetKey: uuidv4(),
-        dataSampleKeys: [],
-        worker: name,
-        metricKey: uuidv4(),
-        metricName: name,
-        computePlanKey: uuidv4(),
-    };
 }
 
 export const getSeriesNodes = (series: SerieT[]): NodeType[] => {
@@ -252,18 +142,6 @@ export const getMaxRankWithPerf = (series: SerieT[]): number => {
     const points = getAllPoints(series);
     const ranks = points.filter((p) => p.perf !== null).map((p) => p.rank);
     return Math.max(0, ...ranks);
-};
-
-export const getMaxEpoch = (series: SerieT[]): number => {
-    const points = getAllPoints(series);
-    const epochs = points.map((p) => p.epoch);
-    return Math.max(0, ...epochs);
-};
-
-export const getMaxEpochWithPerf = (series: SerieT[]): number => {
-    const points = getAllPoints(series);
-    const epochs = points.filter((p) => p.perf !== null).map((p) => p.epoch);
-    return Math.max(0, ...epochs);
 };
 
 export const getMaxRound = (series: SerieT[]): number => {