+++ src/routes/computePlanDetails/components/TasksBreadCrumbs.tsx
@@ -2,7 +2,6 @@ import { BreadcrumbItem, HStack, Text } from '@chakra-ui/react';
 import { RiStackshareLine } from 'react-icons/ri';
 
 import useAppSelector from '@/hooks/useAppSelector';
-import { getMelloddyName } from '@/modules/computePlans/ComputePlanUtils';
 import { PATHS } from '@/routes';
 
 import Breadcrumbs from '@/components/Breadcrumbs';
@@ -33,9 +32,7 @@ const ComputePlanTasksBreadcrumbs = (): JSX.Element => {
                         {computePlanLoading && 'Loading'}
                         {!computePlanLoading &&
                             computePlan &&
-                            (MELLODDY
-                                ? getMelloddyName(computePlan)
-                                : computePlan.tag || 'Untagged compute plan')}
+                            (computePlan.tag || 'Untagged compute plan')}
                     </Text>
                     {!computePlanLoading && computePlan && (
                         <Status